Move DevAddrCacheInfo and IoTHubDeviceInfo to LoraTools
