+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoraTools/DevAddrCacheInfo.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace LoraKeysManagerFacade
+namespace LoRaTools
 {
     using System;
