+++ Tests/Integration/LockDevAddrHelper.cs
@@ -4,6 +4,7 @@
 namespace LoRaWan.Tests.Integration
 {
     using LoraKeysManagerFacade;
+    using LoRaTools;
     using System;
     using System.Threading.Tasks;
